BOL-375 refactor: user 관련 model(in data module) package 분리
