+++ data/src/main/java/com/yapp/bol/data/model/user/JoinedGroupResponse.kt
@@ -1,0 +1,5 @@
+package com.yapp.bol.data.model.user
+
+data class JoinedGroupResponse(
+    val contents: List<JoinedGroupDTO>
+)
