+++ data/src/main/java/com/yapp/bol/data/model/user/OnBoardResponse.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.auth
+package com.yapp.bol.data.model.user
 
 data class OnBoardResponse(
     val onboarding: List<String>
