+++ data/src/main/java/com/yapp/bol/data/model/user/UserRequest.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.auth
+package com.yapp.bol.data.model.user
 
 data class UserRequest(
     val nickname: String
